+++ hutool-json/src/main/java/cn/hutool/json/InternalJSONUtil.java
@@ -1,6 +1,7 @@
 package cn.hutool.json;
 
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.CharUtil;
 import cn.hutool.core.util.NumberUtil;
 import cn.hutool.core.util.ObjectUtil;
@@ -17,7 +18,7 @@ import java.util.SortedMap;
  *
  * @author Looly
  */
-final class InternalJSONUtil {
+public final class InternalJSONUtil {
 
 	private InternalJSONUtil() {
 	}
@@ -28,27 +29,30 @@ final class InternalJSONUtil {
 	 * @param obj 被检查的对象
 	 * @throws JSONException If o is a non-finite number.
 	 */
-	protected static void testValidity(Object obj) throws JSONException {
+	static void testValidity(Object obj) throws JSONException {
 		if (false == ObjectUtil.isValidIfNumber(obj)) {
 			throw new JSONException("JSON does not allow non-finite numbers.");
 		}
 	}
 
 	/**
-	 * 值转为String，用于JSON中。 If the object has an value.toJSONString() method, then that method will be used to produce the JSON text. <br>
-	 * The method is required to produce a strictly conforming text. <br>
-	 * If the object does not contain a toJSONString method (which is the most common case), then a text will be produced by other means. <br>
-	 * If the value is an array or Collection, then a JSONArray will be made from it and its toJSONString method will be called. <br>
-	 * If the value is a MAP, then a JSONObject will be made from it and its toJSONString method will be called. <br>
-	 * Otherwise, the value's toString method will be called, and the result will be quoted.<br>
+	 * 值转为String，用于JSON中。规则为：
+	 * <ul>
+	 *     <li>对象如果实现了{@link JSONString}接口，调用{@link JSONString#toJSONString()}方法</li>
+	 *     <li>对象如果实现了{@link JSONString}接口，调用{@link JSONString#toJSONString()}方法</li>
+	 *     <li>对象如果是数组或Collection，包装为{@link JSONArray}</li>
+	 *     <li>对象如果是Map，包装为{@link JSONObject}</li>
+	 *     <li>对象如果是数字，使用{@link NumberUtil#toStr(Number)}转换为字符串</li>
+	 *     <li>其他情况调用toString并使用双引号包装</li>
+	 * </ul>
 	 *
 	 * @param value 需要转为字符串的对象
 	 * @return 字符串
 	 * @throws JSONException If the value is or contains an invalid number.
 	 */
-	protected static String valueToString(Object value) throws JSONException {
+	static String valueToString(Object value) throws JSONException {
 		if (value == null || value instanceof JSONNull) {
-			return "null";
+			return JSONNull.NULL.toString();
 		}
 		if (value instanceof JSONString) {
 			try {
@@ -66,7 +70,7 @@ final class InternalJSONUtil {
 		} else if (value instanceof Collection) {
 			Collection<?> coll = (Collection<?>) value;
 			return new JSONArray(coll).toString();
-		} else if (value.getClass().isArray()) {
+		} else if (ArrayUtil.isArray(value)) {
 			return new JSONArray(value).toString();
 		} else {
 			return JSONUtil.quote(value.toString());
@@ -79,16 +83,13 @@ final class InternalJSONUtil {
 	 * @param string A String.
 	 * @return A simple JSON value.
 	 */
-	protected static Object stringToValue(String string) {
+	public static Object stringToValue(String string) {
 		// null处理
-		if (null == string || "null".equalsIgnoreCase(string)) {
+		if (StrUtil.isEmpty(string) || StrUtil.NULL.equalsIgnoreCase(string)) {
 			return JSONNull.NULL;
 		}
 
 		// boolean处理
-		if (0 == string.length()) {
-			return StrUtil.EMPTY;
-		}
 		if ("true".equalsIgnoreCase(string)) {
 			return Boolean.TRUE;
 		}
@@ -130,7 +131,7 @@ final class InternalJSONUtil {
 	 * @param value      值
 	 * @return JSONObject
 	 */
-	protected static JSONObject propertyPut(JSONObject jsonObject, Object key, Object value) {
+	static JSONObject propertyPut(JSONObject jsonObject, Object key, Object value) {
 		final String[] path = StrUtil.splitToArray(Convert.toStr(key), CharUtil.DOT);
 		int last = path.length - 1;
 		JSONObject target = jsonObject;
@@ -160,7 +161,7 @@ final class InternalJSONUtil {
 	 * @return 是否忽略null值
 	 * @since 4.3.1
 	 */
-	protected static boolean defaultIgnoreNullValue(Object obj) {
+	static boolean defaultIgnoreNullValue(Object obj) {
 		return (false == (obj instanceof CharSequence))//
 				&& (false == (obj instanceof JSONTokener))//
 				&& (false == (obj instanceof Map));
@@ -178,12 +179,12 @@ final class InternalJSONUtil {
 	 * @return 是否有序
 	 * @since 5.7.0
 	 */
-	protected static boolean isOrder(Object value){
-		if(value instanceof LinkedHashMap || value instanceof SortedMap){
+	static boolean isOrder(Object value) {
+		if (value instanceof LinkedHashMap || value instanceof SortedMap) {
 			return true;
-		} else if(value instanceof JSONGetter){
+		} else if (value instanceof JSONGetter) {
 			final JSONConfig config = ((JSONGetter<?>) value).getConfig();
-			if(null != config){
+			if (null != config) {
 				return config.isOrder();
 			}
 		}