+++ hutool-json/src/main/java/cn/hutool/json/xml/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * JSON与XML相互转换封装，基于json.org官方库改造
+ *
+ * @author looly
+ */
+package cn.hutool.json.xml;